refactor(modulith): extract public browse into discovery module (drop businesses -> servicesoffered edge)
- New discovery module (OPEN) depends on businesses + servicesoffered (leaf, no cycle)
- Move PublicBusinessController/PublicBrowseService/public models out of businesses.publicapi
- Move findPublicQualified* + BusinessWithDistance into DiscoveryBusinessRepository
- businesses core no longer references servicesoffered; route + permitAll (unauthenticated browse) unchanged
- ModularityTests.verify() green; 288 tests
