+++ src/main/kotlin/com/mudhut/nudge/discovery/models/BusinessSort.kt
@@ -1,4 +1,4 @@
-package com.mudhut.nudge.businesses.publicapi.models
+package com.mudhut.nudge.discovery.models
 
 enum class BusinessSort {
     NEWEST,
